refactor: streamline avatar upload and fetching logic
